refactor(ai)!: taxonomy moves, class splits, and cleanups (0.52.0)
videopython.ai technical-debt cleanup (PR-2): the taxonomy moves, file/class splits, and remaining cleanups from the TODO audit. Reorganization, not feature work -- no AI capability changes.

Tier 3 (breaking module moves -- leaf capabilities vs orchestrators): generation/translation.py -> dubbing/translation.py; generation/_tts_backend.py -> dubbing/_tts_backend.py; understanding/separation.py -> dubbing/separation.py (fixes the leaf->orchestrator import inversion; _merge_regions is now public merge_regions).

Tier 4 (breaking class splits): FaceTracker -> FaceSmoothingTracker (single-subject EMA, used by FaceTrackingCrop) + FaceShotTracker (per-shot IoU multi-track, used by analysis), sharing _FaceDetector; AudioClassifier moved from understanding/audio.py to understanding/classification.py (re-exports unchanged; only the deep audio import path changed).

Tier 5 (cleanups): video_analysis/stages.py -> detectors.py; AiError base for the AI error types; public videopython.ai.keyframe (downscale/encode) so mcp/server stops importing ai._ollama privates; dubbing/audio_ops.py merges expressiveness + loudness; TimingAdjustment -> models; DubbingConfig.from_args(); source_metadata.py + try_init extracted from analyzer.py.

Deferred (pure-internal, no API change): pipeline process()/revoice() Demucs dedup; dropping speculative SUPPORTED_MODELS/STEM_NAMES config.

Full AI + editing + base suites pass; ruff + mypy clean.
