+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "videopython"
-version = "0.51.0"
+version = "0.52.0"
 description = "Minimal video generation and processing library."
 authors = [
     { name = "Bartosz Wójtowicz", email = "bartoszwojtowicz@outlook.com" },
