+++ src/videopython/ai/_ollama.py
@@ -2,19 +2,17 @@
 
 from __future__ import annotations
 
-import base64
-import io
 import json
 from typing import Any
 
-import cv2
 import numpy as np
-from PIL import Image
 
 from videopython.ai._optional import require
+from videopython.ai.errors import AiError
+from videopython.ai.keyframe import encode_png_b64
 
 
-class OllamaError(RuntimeError):
+class OllamaError(AiError, RuntimeError):
     """Ollama returned unusable output (non-JSON or an unexpected shape)."""
 
 
@@ -50,7 +48,7 @@ class OllamaStructuredClient:
         """Return the parsed JSON object Ollama generates under ``schema``."""
         user: dict[str, Any] = {"role": "user", "content": text}
         if images:
-            user["images"] = [_encode_png_b64(image) for image in images]
+            user["images"] = [encode_png_b64(image) for image in images]
         messages = [{"role": "system", "content": system}, user]
         response = self._get_client().chat(model=self.model, messages=messages, format=schema, options=self.options)
         content = response.message.content
@@ -64,23 +62,3 @@ class OllamaStructuredClient:
 
     def unload(self) -> None:
         self._client = None
-
-
-# Used only on the MCP keyframe path (videopython.mcp). SceneVLM captioning and the local
-# planner deliberately encode full-resolution frames via _encode_png_b64.
-KEYFRAME_MAX_DIM = 768  # bound a keyframe's longest side before PNG-encoding for the MCP payload
-
-
-def _downscale(frame: np.ndarray, max_dim: int = KEYFRAME_MAX_DIM) -> np.ndarray:
-    """Shrink an RGB frame so its longest side is at most ``max_dim`` (aspect preserved; never upscales)."""
-    h, w = frame.shape[:2]
-    scale = max_dim / max(h, w)
-    if scale >= 1.0:
-        return frame
-    return cv2.resize(frame, (max(1, round(w * scale)), max(1, round(h * scale))), interpolation=cv2.INTER_AREA)
-
-
-def _encode_png_b64(frame: np.ndarray) -> str:
-    buffer = io.BytesIO()
-    Image.fromarray(frame).save(buffer, format="PNG")
-    return base64.b64encode(buffer.getvalue()).decode("ascii")
