+++ src/videopython/ai/understanding/__init__.py
@@ -15,17 +15,19 @@ from videopython.ai._optional import lazy_exports
 if TYPE_CHECKING:
     # Redundant aliases = intentional re-exports (visible to mypy/IDE, not
     # flagged by ruff). Runtime resolution is lazy via __getattr__ below.
-    from .audio import AudioClassifier as AudioClassifier
     from .audio import AudioToText as AudioToText
-    from .faces import FaceTracker as FaceTracker
+    from .classification import AudioClassifier as AudioClassifier
+    from .faces import FaceShotTracker as FaceShotTracker
+    from .faces import FaceSmoothingTracker as FaceSmoothingTracker
     from .image import SceneVLM as SceneVLM
     from .objects import ObjectDetector as ObjectDetector
     from .temporal import SemanticSceneDetector as SemanticSceneDetector
 
 _exports: dict[str, str] = {
     "AudioToText": ".audio",
-    "AudioClassifier": ".audio",
-    "FaceTracker": ".faces",
+    "AudioClassifier": ".classification",
+    "FaceShotTracker": ".faces",
+    "FaceSmoothingTracker": ".faces",
     "ObjectDetector": ".objects",
     "SceneVLM": ".image",
     "SemanticSceneDetector": ".temporal",
